# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API cho hệ thống smart parking",
        "title": "Smart Parking Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/parking-lots": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Trả về toàn bộ danh sách bãi đỗ xe",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parking_lot"
                ],
                "summary": "Lấy danh sách bãi đỗ xe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tạo mới một bãi đỗ xe trong hệ thống",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parking_lot"
                ],
                "summary": "Tạo bãi đỗ xe",
                "parameters": [
                    {
                        "description": "Thông tin bãi đỗ xe",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parking_lot.CreateParkingLotRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/parking-lots/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy thông tin chi tiết của một bãi đỗ xe theo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parking_lot"
                ],
                "summary": "Lấy chi tiết bãi đỗ xe",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID bãi đỗ xe",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cập nhật thông tin bãi đỗ xe theo ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parking_lot"
                ],
                "summary": "Cập nhật bãi đỗ xe",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID bãi đỗ xe",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Thông tin cập nhật bãi đỗ xe",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parking_lot.UpdateParkingLotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/parking-slots/{slotId}/vehicle-logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Trả về lịch sử xe ra vào hoặc hoạt động theo slot ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicle_log"
                ],
                "summary": "Lấy lịch sử xe theo vị trí đỗ",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID vị trí đỗ xe",
                        "name": "slotId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "parking_lot.CreateParkingLotRequest": {
            "type": "object",
            "required": [
                "location",
                "name"
            ],
            "properties": {
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "parking_lot.UpdateParkingLotRequest": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}